+++ packages/tools/atlas/CHANGELOG.md
@@ -1,3 +1,7 @@
+## `@svizzle/atlas` v0.2.0
+
+- moved the distibution dir from a [gist](https://gist.github.com/mindrones/b9538f1b7308d1a2f2d54c927116e825) to the [svizzle_atlas_distro](https://github.com/nestauk/svizzle_atlas_distro) repo
+
 ## `@svizzle/atlas` v0.1.0
 
 - moved world boundaries scripts from /choropleth